requirements.txt
asgiref==3.6.0
certifi==2020.12.5
dj-database-url==1.2.0
Django==4.1.5
djangorestframework==3.13.1
elasticsearch==7.12.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==6.3.0
